feat: HiddenTextarea, IconButtons 코드 분리
