+++ src/pages/home/subPages/main/components/MemoCreateTextArea/HiddenTextarea/index.ts
@@ -1,0 +1,1 @@
+export { default as HiddenTextarea } from './HiddenTextarea';
