+++ src/pages/home/subPages/main/components/MemoCreateTextArea/IconButtons/index.ts
@@ -1,0 +1,1 @@
+export { default as IconButtons } from './IconButtons';
